+++ website/documentation-site.ts
@@ -21,6 +21,7 @@ export const DOCUMENTATION_DEFAULT_VERSION = 'latest';
 
 export const DOCUMENTATION_VERSIONS = [
   'latest',
+  '0.307.0',
   '0.306.0',
   '0.305.0',
   '0.304.0',
@@ -40,7 +41,6 @@ export const DOCUMENTATION_VERSIONS = [
   '0.290.0',
   '0.289.0',
   '0.288.0',
-  '0.287.0',
   '0.38.0',
   '0.1.0',
 ] as const;